backend/app/core/security.py
from datetime import datetime, timedelta
from jose import jwt
from app.core.config import settings


ALGO = "HS256"


def create_access_token(sub: str) -> str:
    expire = datetime.utcnow() + timedelta(minutes=settings.JWT_EXPIRE_MINUTES)
    to_encode = {"sub": sub, "exp": expire}
    return jwt.encode(to_encode, settings.JWT_SECRET, algorithm=ALGO)