backend/pyproject.toml
[tool.poetry]
name = "pulseboard"
version = "0.1.0"
description = "Pulseboard FastAPI backend"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.0"
uvicorn = {extras=["standard"], version="^0.30.0"}
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
sqlalchemy = "^2.0.31"
psycopg = {version="^3.1.19", extras=["binary"]}
python-jose = "^3.3.0"
passlib = {version = "^1.7.4", extras=["bcrypt"]}
redis = "^5.0.4"
structlog = "^24.1.0"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.8"
ruff = "^0.5.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
alembic = "^1.13.1"


[tool.black]
line-length = 100


[tool.ruff]
line-length = 100